Split bootstrap code so it doesn't depend on `package:analyzer`.
