+++ build_runner/lib/src/bootstrap/build_script_generate.dart
@@ -5,7 +5,6 @@
 import 'dart:async';
 
 import 'package:build_config/build_config.dart';
-import 'package:dart_style/dart_style.dart';
 import 'package:path/path.dart' as p;
 import 'package:pub_semver/pub_semver.dart';
 
@@ -13,29 +12,25 @@ import '../build_plan/build_configs.dart';
 import '../build_plan/build_packages.dart';
 import '../build_plan/build_paths.dart';
 import '../constants.dart';
-import '../exceptions.dart';
 import '../io/reader_writer.dart';
-import '../logging/build_log.dart';
 
 final _lastShortFormatDartVersion = Version(3, 6, 0);
 
 Future<String> generateBuildScript({required BuildPaths buildPaths}) async {
   final builderFactories = await loadBuilderFactories(buildPaths: buildPaths);
 
-  try {
-    // The `build_runner` version number is included to force a rebuild if the
-    // script was generated by a different version. It only needs updating if
-    // the host<->isolate relationship changed in a breaking way, for example
-    // if command line args or messages passed via sendports have changed
-    // in a breaking way.
-    return DartFormatter(languageVersion: _lastShortFormatDartVersion).format(
-      '''
+  // The `build_runner` version number is included to force a rebuild if the
+  // script was generated by a different version. It only needs updating if
+  // the host<->isolate relationship changed in a breaking way, for example
+  // if command line args or messages passed via sendports have changed
+  // in a breaking way.
+  return '''
 // @dart=${_lastShortFormatDartVersion.major}.${_lastShortFormatDartVersion.minor}
 // ignore_for_file: type=lint
 // build_runner >=2.4.16
 import 'dart:io' as _io;
 import 'package:build_runner/src/build_plan/builder_factories.dart' as _build_runner;
-import 'package:build_runner/src/bootstrap/processes.dart' as _build_runner;
+import 'package:build_runner/src/bootstrap/child_process.dart' as _build_runner;
 ${builderFactories.renderImports}
 
 final _builderFactories = _build_runner.BuilderFactories(
@@ -48,15 +43,7 @@ void main(List<String> args) async {
     _builderFactories,
   )!;
 }
-''',
-    );
-  } on FormatterException {
-    buildLog.error(
-      'Generated build script could not be parsed. '
-      'Check builder definitions.',
-    );
-    throw const CannotBuildException();
-  }
+''';
 }
 
 /// Loads builder factories for the current package and its transitive
